fix: address Claude Review findings on analytics PR
- Add `sprints` field to TreeResponse interface in api.ts
- Add missing `sprints: []` to audit and progress test fixtures
- Use null-safe checks (== null) instead of falsy for graph edge coords
- Extract shared `progressBar` to cli/utils/output.ts, removing
  duplication between status.ts and sprint.ts
